refactor: move DIRAC TS WebApp related RPC calls to dedicated service
